% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_devices}
\alias{check_devices}
\title{Looks for serial ports in use.}
\usage{
check_devices()
}
\value{
A character vector with all active devices, or `character(0)` if no device is found.
}
\description{
This function checks for serial ports in use on the host device.
}
\examples{
\dontrun{
available_devices <- check_devices()
}
}

