% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial.R
\name{close.serialConnection}
\alias{close}
\alias{close.serialConnection}
\title{Function to close a serial interface.}
\usage{
\method{close}{serialConnection}(con, ...)
}
\arguments{
\item{con}{Object of class \code{serialConnection}.}

\item{...}{Currently ignored.}
}
\description{
This function closes the corresponding connection.
}
\examples{
\dontrun{
close(con)
}

}
\seealso{
\code{\link{serialConnection}} \code{\link{open}}
}

