% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{FindDevices}
\alias{FindDevices}
\title{Find serial ports in use.}
\usage{
FindDevices()
}
\value{
A vector of strings containing the names of all active devices, or 
`character(0)` if no device is found.
}
\description{
This function checks for serial ports in use on the host device.
}
\examples{
\dontrun{
# Use the first active device.
available_devices <- FindDevices()

con <- serialConnection(available_devices[1])
}
}

