% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial.R
\name{flush.serialConnection}
\alias{flush.serialConnection}
\title{Flushes the connection.}
\usage{
\method{flush}{serialConnection}(con)
}
\arguments{
\item{con}{serial connection}
}
\value{
Nothing is returned
}
\description{
Some times (and depending on buffering mode) the connection buffer needs 
to be flushed manually. This command empties the buffer by sending all 
remaining bytes.
}
\examples{
 # See the top package documentation
}
\seealso{
\code{\link{serial}}
}
