% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{isOpen.default}
\alias{isOpen.default}
\title{Default function from base-package}
\usage{
\method{isOpen}{default}(con, rw = "")
}
\arguments{
\item{con}{connection object}

\item{rw}{defines the mode of operation}
}
\description{
Default function from base-package
}
\seealso{
\code{\link[base]{isOpen}}
}

