% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{isOpen.serialConnection}
\alias{isOpen.serialConnection}
\title{Tests whether the connection is open or not}
\usage{
\method{isOpen}{serialConnection}(con, ...)
}
\arguments{
\item{con}{connection of the class \code{serialConnection}}

\item{...}{not used}
}
\value{
returns \code{{F, T}} for 'not open' and 'is open'
}
\description{
Tests whether the connection is open or not
}

