% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial.R
\name{nBytesInQueue}
\alias{nBytesInQueue}
\title{Reads out the number of characters / bytes pending in the input and output buffer}
\usage{
nBytesInQueue(con)
}
\arguments{
\item{con}{serial connection}
}
\value{
named vector with number of bytes
}
\description{
Reads out the number of characters / bytes pending in the input and output buffer
}
\examples{
 # See the top package documentation
}
\seealso{
\code{\link{serial}}
}
