% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial.R
\name{open}
\alias{open}
\alias{open.serialConnection}
\title{Function to initialize an serial interface.}
\usage{
\method{open}{serialConnection}(con, ...)
}
\arguments{
\item{con}{Object of class \code{serialConnection}.}

\item{...}{Currently ignored.}
}
\description{
This function initializes the serial interface and opens it for later usage.
}
\examples{
\dontrun{
con <- serialConnection(port = 'com5')
open(con)
}
}
\seealso{
\code{\link{serialConnection}} \code{\link{close}}
}

