% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{open.serialConnection}
\alias{open.serialConnection}
\title{Function to initialize an serial interface.}
\usage{
\method{open}{serialConnection}(con, ...)
}
\arguments{
\item{con}{serial connection}

\item{...}{is ignored}
}
\description{
This function initializes the serial interface and opens it for later usage.
}
\seealso{
\code{\link{serialConnection}}
}

