% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.serialConnection}
\alias{print.serialConnection}
\title{Print method for serialConnection}
\usage{
\method{print}{serialConnection}(x, ...)
}
\arguments{
\item{x}{\code{serialConnection} Object}

\item{...}{not used}
}
\value{
nothing to return
}
\description{
Outputs major information of the serial connection.
}
