% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read.serialConnection}
\alias{read.serialConnection}
\title{Writes data to the serial interface.}
\usage{
read.serialConnection(con, ...)
}
\arguments{
\item{con}{Object of class \code{serialConnection}}

\item{...}{Currently ignored.}
}
\description{
Deprecated, please use \code{\link{read_data}}.
}

