% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{read_data}
\alias{read_data}
\alias{read_data.serialConnection}
\title{Reads from the serial interface.}
\usage{
read_data(con, ...)
}
\arguments{
\item{con}{serial connection}

\item{...}{unused for now.}
}
\value{
The result is a string, which can be converted to raw as necessary
}
\description{
This function reads from the serial interface as long as the buffer is not
empty. The read takes place per byte.
}
\examples{
 # See the top package documentation
}
\seealso{
\code{\link{serial}}
}

