% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{serial}
\alias{serial}
\alias{serial-package}
\title{A serial communication interface for R.}
\description{
\describe{
 \item{\code{\link{serialConnection}}}{Creates a \code{serialConnection} object}
 \item{\code{\link{open}}}{opens a serial connection}
 \item{\code{\link{close}}}{closes the serial connection}
 \item{\code{\link{read_data}}}{bytewise read from the interface as long as the buffer is empty}
 \item{\code{\link{write_data}}}{writes a string to the serial interface}
 \item{\code{\link{find_devices}}}{helper function to find available serial devices}
}
}
\details{
This R package provides the functionality to use the serial communication ports
"COM" to use RS232/RS422/RS485 functionality of the corresponding hardware.
Also, virtual COM ports via USB do work, as long as they are mapped to COM[n] (Windows)
or tty[n] (Mac/Linux) in the operating system.
}
\examples{
# For this example, I used the 'null-modem' emulator 'com0com' for Windows
# which is available on 'http://com0com.sourceforge.net/'.
# Here, the pair of COM ports is 'CNCA0' <-> 'CNCB0'.

# Test the functionality:
# ======================
#
# First: Install the virtual null-modem connection like
#        com0com (win) or tty0tty (linux)
#        Hint: Some unix insist on port names like 'ttxS[n]'.
# 
# Second: Set up a terminal program (like HTerm or gtkterm) and listen to 
#         COM port 'CNCB0' (or what ever you have installed)
#         or (for unix only) 'cat /dev/tnt1' will output tnt1 to the console.

\dontrun{

# Now configure one of the COM ports with appropriate connection properties
con <- serialConnection(port = "CNCA0", # or "tnt0" on linux
                        mode = "115200,n,8,1",
                        newline = 1,
                        translation = "crlf")

# Let's open the serial interface.

open(con)

# write some stuff
write_data(con, "Hello World!")

# read, in case something came in
read_data(con)

# close the connection
close(con)
}
}
\concept{
RS232 serial

RS485 serial

RS422 serial

USB serial
}

