% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.serialConnection}
\alias{summary.serialConnection}
\title{Serial Connection Summary}
\usage{
\method{summary}{serialConnection}(object, ...)
}
\arguments{
\item{object}{object of type \code{serialConnection}}

\item{...}{not used}
}
\value{
Table of connection properties
}
\description{
Displays summarized informations and status of the serial connection.
}
