% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write_data}
\alias{write_data}
\alias{write_data.serialConnection}
\title{Writes data to serial interface.}
\usage{
write_data(con, dat, ...)
}
\arguments{
\item{con}{serial connection}

\item{dat}{data string to write on the serial interface. This can either
be a character string or any R object that can be coerced
to a character string by the \code{\link{toString}} function,
though the latter occurs with a warning. 
See the example section in \code{\link{serial}}.}

\item{...}{unused for now.}
}
\description{
Writes data to serial interface.
}
\examples{
 # See the top package documentation
 
 \dontrun{write_data(con, "Hello World!")}
}
\seealso{
\code{\link{serial}}
}

