% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial.R
\name{check_devices}
\alias{check_devices}
\title{Looks for active serial ports.}
\usage{
check_devices()
}
\value{
A character vector with all active devices, or `character(0)` if no device is found.
}
\description{
This function checks for active serial ports on the host device.
}

