% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{close.serialConnection}
\alias{close.serialConnection}
\title{Function to close an serial interface.}
\usage{
\method{close}{serialConnection}(con, ...)
}
\arguments{
\item{con}{serial connection}

\item{...}{is ignored}
}
\description{
This function closes the corresponding connection.
}
\seealso{
\code{\link{serialConnection}}
}

