% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{find_devices}
\alias{find_devices}
\title{Looks for serial ports in use.}
\usage{
find_devices()
}
\value{
A vector of character elements containing all active devices, or 
`character(0)` if no device is found.
}
\description{
This function checks for serial ports in use on the host device.
}
\examples{
\dontrun{
available_devices <- find_devices()
}
}

