% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial.R
\name{read.serialConnection}
\alias{read.serialConnection}
\title{Reads from the serial interface.}
\usage{
read.serialConnection(con)
}
\arguments{
\item{con}{serial connection}
}
\value{
The result is a string, which can be converted to raw as necessary
}
\description{
This function reads from the serial interface as long as the buffer is not
empty. The read takes place per byte.
}
\examples{
 # See the top package documentation
}
\seealso{
\code{\link{serial}}
}

