% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{write.serialConnection}
\alias{write.serialConnection}
\title{Writes data to the serial interface.}
\usage{
write.serialConnection(con, dat, ...)
}
\arguments{
\item{con}{Object of class \code{serialConnection}}

\item{dat}{Data string to write on the serial interface. This can either
be a character string or any R object that can be coerced
to a character string by the \code{\link{toString}} function,
though the latter occurs with a warning. 
See the example section in \code{\link{serial}}.}

\item{...}{Currently ignored.}
}
\description{
Deprecated, please use \code{\link{write_data}}.
}

