% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial.R
\name{write.serialConnection}
\alias{write.serialConnection}
\title{Wirtes data to serial interface.}
\usage{
write.serialConnection(con,dat)
}
\arguments{
\item{con}{serial Connection}

\item{dat}{data string to write on the serial interface. At the moment this 
must be a string \code{'...'}. See examle section in \code{\link{serial}}.}
}
\description{
Wirtes data to serial interface.
}
\examples{
 # See the top package documentation
 
 \dontrun{write.serialConnection(con, 'Hello World!')}
}
\seealso{
\code{\link{serial}}
}

