% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serial.R
\name{write_data}
\alias{write.serialConnection}
\alias{write_data}
\alias{write_data.serialConnection}
\title{Writes data to the serial interface.}
\usage{
\method{write_data}{serialConnection}(con, dat, ...)
}
\arguments{
\item{con}{Object of class \code{serialConnection}}

\item{dat}{Data string to write on the serial interface. This must be a character string.}

\item{...}{Currently ignored.}
}
\description{
Writes data to the serial interface.
}
\examples{
# See `help(serial)` for a complete example.
 
\dontrun{
write_data(con, "Hello World!")
write_data(con, 2) # error
}
}
\seealso{
\code{\link{serial}} \code{\link{read_data}}
}

