% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/serial.R
\name{write_data}
\alias{write.serialConnection}
\alias{write_data}
\alias{write_data.serialConnection}
\title{Writes data to the serial interface.}
\usage{
write_data(con, dat, ...)

\method{write_data}{serialConnection}(con, dat, ...)
}
\arguments{
\item{con}{Object of class \code{serialConnection}}

\item{dat}{Data string to write on the serial interface. This can either
be a character string or any R object that can be coerced
to a character string by the \code{\link{toString}} function,
though the latter occurs with a warning. 
See the example section in \code{\link{serial}}.}

\item{...}{Currently ignored.}
}
\description{
Writes data to the serial interface.
}
\examples{
# See the top package documentation for a complete example.
 
\dontrun{
write_data(con, "Hello World!")
write_data(con, 2) # coerce with warning
}
}
\seealso{
\code{\link{serial}} \code{\link{read_data}}
}

